programs.sway: always wrap the provided package
when swapping the default package with another (e.g., "swayfx") we lose
all kinds of HM properties like the session variable / extra options etc
for no reason. It's inconsistent with how the neovim module behaves too.

If someone wants to bypass the HM wrapping, it's still possible to wrap
the package with a function .

programs.sway: always wrap the provided package

when swapping the default package with another (e.g., "swayfx") we lose
all kinds of HM properties like the session variable / extra options etc
for no reason. It's inconsistent with how the neovim module behaves too.

If someone wants to bypass the HM wrapping, it's still possible to wrap
the package with a function .

ci: attempts to fix